Serialise manually to work around serde bugs.
It seems that various serde libraries are trying to decode
Option<RistrettoPoint>s as RistrettoPoints, which then causes the
`impl Deserialize for RistrettoPoint` implementation to error because the
Option<T> has a length of one byte.

I tried with both bincode and serde_cbor.  Rather than find the bug, it's
probably faster to write my own serialisation code manually.